+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/navigator/AppNavigator.kt
@@ -94,6 +94,10 @@ class AppNavigator @Inject constructor() : Navigator{
         navigate(OrganismsFragmentDirections.actionOrganismsFragmentToPrimaryCardViewFragment())
     }
 
+    override fun Fragment.goToInformationMessageCardView() {
+        navigate(OrganismsFragmentDirections.actionOrganismsFragmentToInformationMessageCardViewFragment())
+    }
+
     override fun Fragment.goBack() {
         findNavController().popBackStack()
     }