+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/navigator/Navigator.kt
@@ -61,5 +61,7 @@ interface Navigator {
 
     fun Fragment.goToPrimaryCardView()
 
+    fun Fragment.goToInformationMessageCardView()
+
     fun Fragment.goBack()
 }