+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/presentation/fragment/organisms/OrganismsFragment.kt
@@ -28,6 +28,7 @@ import uk.gov.hmrc.sample_compose_components.R
 import uk.gov.hmrc.sample_compose_components.databinding.FragmentComposeExampleBinding
 import uk.gov.hmrc.sample_compose_fragments.data.repository.RepositoryImpl.Companion.ORGANISM_HEADLINE_CARD_VIEW
 import uk.gov.hmrc.sample_compose_fragments.data.repository.RepositoryImpl.Companion.ORGANISM_ICON_BUTTON_CARD_VIEW
+import uk.gov.hmrc.sample_compose_fragments.data.repository.RepositoryImpl.Companion.ORGANISM_INFORMATION_MESSAGE_CARD_VIEW
 import uk.gov.hmrc.sample_compose_fragments.data.repository.RepositoryImpl.Companion.ORGANISM_PRIMARY_CARD_VIEW
 import uk.gov.hmrc.sample_compose_fragments.data.repository.RepositoryImpl.Companion.ORGANISM_SEPARATED_VIEW_CONTAINER
 import uk.gov.hmrc.sample_compose_fragments.navigator.Navigator
@@ -70,6 +71,10 @@ class OrganismsFragment : Fragment(R.layout.fragment_compose_example) {
 
                                 ORGANISM_PRIMARY_CARD_VIEW -> {
                                     with(navigator) { goToPrimaryCardView() }
+                                }
+
+                                ORGANISM_INFORMATION_MESSAGE_CARD_VIEW -> {
+                                    with(navigator) { goToInformationMessageCardView() }
                                 }
                             }
                         })